refactor:useDrawing 커스텀 훅 관심사 에 맞게 분리
* useDomContentsLoaded 커스텀 훅 분리
* useDrawing 삭제 -> useDrawUtils, useDrawInstance 커스텀 훅 분리
* convertDataToCanvasImageData 메서드 분리
   - utils/ 디렉토리 생성
* draw, erase 함수 위치 변경
