+++ frontend/src/hooks/useDomContentsLoaded.jsx
@@ -1,0 +1,12 @@
+import { useState } from 'react';
+
+const useDomContentsLoaded = () => {
+  const [contentsLoaded, setContentsLoaded] = useState(false);
+  document.addEventListener('DOMContentLoaded', () => {
+    setContentsLoaded(true);
+  });
+
+  return contentsLoaded;
+};
+
+export default useDomContentsLoaded;
